refactor: Member - Group 도메인 분리
- Group 자체적으로 분기되는 플로우가 많아서 분리하는게 제어하기 더 편하다고 판단
