+++ src/main/java/ac/dnd/mur/server/group/domain/model/Group.java
@@ -1,18 +1,16 @@
-package ac.dnd.mur.server.member.domain.model;
+package ac.dnd.mur.server.group.domain.model;
 
+import ac.dnd.mur.server.member.domain.model.Member;
 import ac.dnd.mur.server.member.exception.MemberException;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import static ac.dnd.mur.server.member.exception.MemberExceptionCode.MEMBER_GROUP_NAME_TOO_LONG;
-import static jakarta.persistence.FetchType.LAZY;
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
@@ -27,15 +25,14 @@ public class Group {
     @GeneratedValue(strategy = IDENTITY)
     private Long id;
 
+    @Column(name = "member_id", nullable = false)
+    private Long memberId;
+
     @Column(name = "name", nullable = false)
     private String name;
 
-    @ManyToOne(fetch = LAZY, optional = false)
-    @JoinColumn(name = "member_id", referencedColumnName = "id", nullable = false)
-    private Member member;
-
     private Group(final Member member, final String name) {
-        this.member = member;
+        this.memberId = member.getId();
         this.name = name;
     }
